#41 ActiveCheatkeyAdapter 와 ActiveCheatkeyData 분리
