+++ jibcon/src/main/java/com/sm_arts/jibcon/app/cheatkey/ActiveCheatkeyAdapter.java
@@ -74,16 +74,3 @@ public class ActiveCheatkeyAdapter
         return mActiveCheatkeyDataset.size();
     }
 }
-
-// input 텍스트뷰와 output 텍스트뷰 DataSet
-class ActiveCheatkeyData{
-    public int imgBtnID;
-    public int activeCheatkeyImg;
-    public String activeCheatkeyName;
-
-    public ActiveCheatkeyData(int imgBtnID, int activeCheatkeyImg, String activeCheatkeyName){
-        this.imgBtnID = imgBtnID;
-        this.activeCheatkeyImg = activeCheatkeyImg;
-        this.activeCheatkeyName = activeCheatkeyName;
-    }
-}
