+++ jibcon/src/main/java/com/sm_arts/jibcon/app/cheatkey/ActiveCheatkeyData.java
@@ -1,0 +1,17 @@
+package com.sm_arts.jibcon.app.cheatkey;
+
+/**
+ * Created by woojinkim on 2017. 7. 17..
+ */
+
+public class ActiveCheatkeyData{
+    public int imgBtnID;
+    public int activeCheatkeyImg;
+    public String activeCheatkeyName;
+
+    public ActiveCheatkeyData(int imgBtnID, int activeCheatkeyImg, String activeCheatkeyName){
+        this.imgBtnID = imgBtnID;
+        this.activeCheatkeyImg = activeCheatkeyImg;
+        this.activeCheatkeyName = activeCheatkeyName;
+    }
+}
